Added some tests
The tests that have been currently setup make use of a CARMIN/URL simulator.
This enables the testing of both the update methods found in DescriptorEntry and CarminPlatformEntry and the validation/submission done with DescriptorDataCandidate and CarminPlatformCandidate.
The tests folder also includes sample descriptors including test-cases in /boutiques_descriptors/. 
Those descriptors are automatically loaded by common,py. This ease the referencing of descriptor data.
Among other things, the file tools.py also defines utilities that can be used to perform certain assertions on the database objects.